add new python ai plugin
