+++ PythonAiPlugin/.gitignore
@@ -5,3 +5,4 @@
 /src/main/java/io/antmedia/app/__pycache__/
 /dist/
 /.idea/
+/pythonAIPlugin.zip
