+++ PythonAiPlugin/copy_files.sh
@@ -8,4 +8,11 @@ cp ./web/samples/* $AMS_DIR/webapps/LiveApp/samples
 cp ./web/samples/* $AMS_DIR/webapps/WebRTCAppEE/samples
 cp ./web/samples/* $AMS_DIR/webapps/live/samples
 
-#./start.sh
+mkdir -p PythonPluginFiles 
+cp -r ./src/main/java/io/antmedia/samples  "$AMS_DIR/PythonPluginFiles"
+cp -r ./src/main/java/io/antmedia/samples/init_plugins.py  "$AMS_DIR/PythonPluginFiles"
+cp ./src/main/java/io/antmedia/app/*.py   "$AMS_DIR/PythonPluginFiles"
+
+cd $AMS_DIR
+
+./start.sh
